[MINOR] improvement: fix typos of guava imports (#9431)
### What changes were proposed in this pull request?

Some `Sets` in the project are pointed to the shadowed one in Glassfish,
rather than original Guava.

1. Search through the project to replace all `import
org.glassfish.jersey.internal.guava.Xxss` to its corresponding
`com.google.common.collect.Xxxs`.
2. Add 3 `replaceRegex` actions to Spotless plugin, to ensure no similar
issues in the future.

### Why are the changes needed?

For sure the one from original Guava should be preferred, because
shadowed classes are normally used for internal purposes of the party
who package them, and are not intended to expose as a contract.

Fix: No functionality issue involved. 

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Unit tests all passed. Plus, the implementations of two classes are
identical, so in fact as long as it compiles right, we are good to go I
believe.

---------

Co-authored-by: wancheng.aliyun <tuohai@alibaba-inc.com>